import CircuitNodeId, CircuitEdgeId from utils
