+++ tests/test_circuit_ids.py
@@ -9,39 +9,13 @@ import pytest
 
 import bluepysnap.circuit_ids as test_module
 from bluepysnap.exceptions import BluepySnapError
-from bluepysnap.utils import IDS_DTYPE
+from bluepysnap.utils import IDS_DTYPE, CircuitEdgeId, CircuitNodeId
 
 from utils import setup_tempdir
 
 
 def _multi_index():
     return pd.MultiIndex.from_arrays([["a", "a", "b", "a"], [0, 1, 0, 2]])
-
-
-class TestCircuitNodeId:
-    def setup_method(self):
-        self.test_obj = test_module.CircuitNodeId("pop", 1)
-
-    def test_init(self):
-        assert isinstance(self.test_obj, test_module.CircuitNodeId)
-        assert isinstance(self.test_obj, tuple)
-
-    def test_accessors(self):
-        assert self.test_obj.population == "pop"
-        assert self.test_obj.id == 1
-
-
-class TestCircuitEdgeId:
-    def setup_method(self):
-        self.test_obj = test_module.CircuitEdgeId("pop", 1)
-
-    def test_init(self):
-        assert isinstance(self.test_obj, test_module.CircuitEdgeId)
-        assert isinstance(self.test_obj, tuple)
-
-    def test_accessors(self):
-        assert self.test_obj.population == "pop"
-        assert self.test_obj.id == 1
 
 
 class TestCircuitNodeIds:
@@ -51,7 +25,7 @@ class TestCircuitNodeIds:
 
     @property
     def id_cls(self):
-        return test_module.CircuitNodeId
+        return CircuitNodeId
 
     @property
     def id_name(self):
@@ -335,7 +309,7 @@ class TestCircuitEdgeIds(TestCircuitNodeIds):
 
     @property
     def id_cls(self):
-        return test_module.CircuitEdgeId
+        return CircuitEdgeId
 
     @property
     def id_name(self):
