+++ tests/test_edges.py
@@ -11,9 +11,9 @@ from numpy import dtype
 import bluepysnap.edges as test_module
 from bluepysnap.bbp import Synapse
 from bluepysnap.circuit import Circuit
-from bluepysnap.circuit_ids import CircuitEdgeId, CircuitEdgeIds, CircuitNodeId, CircuitNodeIds
+from bluepysnap.circuit_ids import CircuitEdgeIds, CircuitNodeIds
 from bluepysnap.exceptions import BluepySnapError
-from bluepysnap.utils import IDS_DTYPE
+from bluepysnap.utils import IDS_DTYPE, CircuitEdgeId, CircuitNodeId
 
 from utils import PICKLED_SIZE_ADJUSTMENT, TEST_DATA_DIR
