+++ tests/test_frame_report.py
@@ -8,10 +8,10 @@ import pytest
 
 import bluepysnap.frame_report as test_module
 from bluepysnap.bbp import Cell
-from bluepysnap.circuit_ids import CircuitNodeId, CircuitNodeIds
+from bluepysnap.circuit_ids import CircuitNodeIds
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.simulation import Simulation
-from bluepysnap.utils import IDS_DTYPE
+from bluepysnap.utils import IDS_DTYPE, CircuitNodeId
 
 from utils import TEST_DATA_DIR
