+++ tests/test_spike_report.py
@@ -10,10 +10,10 @@ from libsonata import SpikeReader
 
 import bluepysnap.spike_report as test_module
 from bluepysnap.bbp import Cell
-from bluepysnap.circuit_ids import CircuitNodeId, CircuitNodeIds
+from bluepysnap.circuit_ids import CircuitNodeIds
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.simulation import Simulation
-from bluepysnap.utils import IDS_DTYPE
+from bluepysnap.utils import IDS_DTYPE, CircuitNodeId
 
 from utils import TEST_DATA_DIR, copy_test_data, edit_config
